requirements.txt
asgiref==3.5.0
astroid==2.3.3
bleach==5.0.1
Brotli==1.0.9
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cryptography==36.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-markdownify==0.9.2
django-model-utils==4.2.0
django-resized==0.3.11
django-rest-auth==0.9.5
djangorestframework==3.13.1
fonttools==4.38.0
google-auth==2.14.1
google-auth-oauthlib==0.7.1
gspread==5.5.0
hashids==1.3.1
html5lib==1.1
httplib2==0.21.0
idna==2.8
importlib-metadata==5.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.4.1
mccabe==0.6.1
oauth2client==2.2.0
oauthlib==3.2.0
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.5.0
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyphen==0.13.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2022.1
reportlab==3.6.6
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.13.0
sqlparse==0.4.2
sslcommerz-lib==1.0
tinycss2==1.1.1
typed-ast==1.4.0
typing_extensions==4.1.1
urllib3==1.25.7
weasyprint==57.1
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
zipp==3.10.0
zopfli==0.2.1
